api/requirements.txt
Flask==2.3.3
flask-cors
requests
python-dotenv
pymongo